requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.4.0
    # via
    #   outcome
    #   trio
certifi==2022.6.15
    # via urllib3
cffi==1.15.1
    # via cryptography
cryptography==37.0.4
    # via
    #   pyopenssl
    #   urllib3
h11==0.13.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
outcome==1.2.0
    # via trio
pycparser==2.21
    # via cffi
pymysql==1.0.2
    # via -r requirements.in
pyopenssl==22.0.0
    # via urllib3
pysocks==1.7.1
    # via urllib3
python-dotenv==0.20.0
    # via -r requirements.in
selenium==4.3.0
    # via -r requirements.in
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
trio==0.21.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure,socks]==1.26.10
    # via selenium
wsproto==1.1.0
    # via trio-websocket
